+++ src/zos_console/zowe/zos_console_for_zowe_sdk/console.py
@@ -25,10 +25,12 @@ class Console(SdkApi):
     ----------
     connection : dict
        A profile in dict (json) format
+    log : bool
+        Flag to disable logger
     """
 
-    def __init__(self, connection: dict):
-        super().__init__(connection, "/zosmf/restconsoles/consoles/defcn", logger_name=__name__)
+    def __init__(self, connection: dict, log: bool = True):
+        super().__init__(connection, "/zosmf/restconsoles/consoles/defcn", logger_name=__name__, log=log)
 
     def issue_command(self, command: str, console: Optional[str] = None) -> IssueCommandResponse:
         """Issues a command on z/OS Console.