+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 All notable changes to the Zowe Client Python SDK will be documented in this file.
 
+## Recent Changes
+
+### Enhancements
+
+- Turning of logger at the class-constructor level [#316] (https://github.com/zowe/zowe-client-python-sdk/issues/316)
+
 ## `1.0.0-dev21`
 
 ### Bug Fixes
 
 - Fixed Core SDK package referencing a non-existent version of Secrets SDK.
+
 
 ## `1.0.0-dev20`